+++ frontend/components/feedback-task/header/sort-filter/SortSelector.vue
@@ -12,14 +12,12 @@
       "
     />
     <BaseDropdown
-      :visible="visibleDropdown"
-      @visibility="onToggleVisibility"
-      class="sort-selector__dropdown"
       v-if="nonSelectedSortingItems.length"
+      :visible="visibleDropdown"
     >
       <span slot="dropdown-header">
-        <BaseButton class="secondary small light sort-selector__add-button"
-          >+ Add another field</BaseButton
+        <BaseButton class="secondary small light" @click="onToggleVisibility">
+          + Add another field</BaseButton
         >
       </span>
       <span slot="dropdown-content">
@@ -29,11 +27,6 @@
         ></SortCategoriesList>
       </span>
     </BaseDropdown>
-    <BaseButton
-      class="primary small full-width sort-selector__button"
-      @on-click="applySorting"
-      >Sort</BaseButton
-    >
   </div>
 </template>
 <script>
@@ -58,8 +51,8 @@ export default {
     },
   },
   methods: {
-    onToggleVisibility(value) {
-      this.visibleDropdown = value;
+    onToggleVisibility() {
+      this.visibleDropdown = !this.visibleDropdown;
     },
     includeSortCategory(categoryName) {
       this.sortingItems.select(categoryName);
@@ -92,11 +85,5 @@ export default {
   flex-direction: column;
   gap: $base-space;
   padding: $base-space;
-  &__button {
-    margin-top: $base-space * 2;
-  }
-  &__dropdown {
-    margin-left: 36px;
-  }
 }
 </style>