+++ frontend/v1/domain/entities/metadata/MetadataSort.ts
@@ -1,22 +1,47 @@
 import { Metadata } from "./Metadata";
 
-class MetadataSort {
-  public selected = false;
-  public sort: "asc" | "desc" = "asc";
-  constructor(private metadata: Metadata | { name: string }) {}
+const SORT_KEY_SEPARATOR = ".";
+const ORDER_BY_SEPARATOR = ":";
+const SORT_ASC = "asc";
+const SORT_DESC = "desc";
+
+type SortOptions = "asc" | "desc";
+
+abstract class Sort {
+  public sort: SortOptions = SORT_ASC;
+
+  constructor(public readonly key: string = "") {}
+
+  toggleSort() {
+    this.sort = this.sort === SORT_ASC ? SORT_DESC : SORT_ASC;
+  }
+
+  abstract get name(): string;
+}
+
+class MetadataSort extends Sort {
+  constructor(private metadata: Metadata) {
+    super(`metadata${SORT_KEY_SEPARATOR}`);
+  }
 
   get name() {
     return this.metadata.name;
   }
 }
 
+class RecordSort extends Sort {
+  constructor(public readonly name: string) {
+    super();
+  }
+}
+
 export class MetadataSortList {
-  private metadataSorts: MetadataSort[];
-  private selectedCategories: MetadataSort[] = [];
+  private metadataSorts: Sort[];
+  private selectedCategories: Sort[] = [];
   constructor(metadata: Metadata[] = []) {
     this.metadataSorts = metadata.map((metadata) => new MetadataSort(metadata));
-    this.metadataSorts.push(new MetadataSort({ name: "inserted_at" }));
-    this.metadataSorts.push(new MetadataSort({ name: "updated_at" }));
+    this.metadataSorts.push(new RecordSort("inserted_at"));
+    this.metadataSorts.push(new RecordSort("updated_at"));
   }
 
   get selected() {
@@ -24,7 +49,9 @@ export class MetadataSortList {
   }
 
   get noSelected() {
-    return this.metadataSorts.filter((metadataSort) => !metadataSort.selected);
+    return this.metadataSorts.filter(
+      (metadata) => !this.selectedCategories.includes(metadata)
+    );
   }
 
   get selectedCategoriesName() {
@@ -34,7 +61,6 @@ export class MetadataSortList {
   select(category: string) {
     const found = this.findByCategory(category);
     if (found) {
-      found.selected = true;
       this.selectedCategories.push(found);
     }
   }
@@ -45,47 +71,53 @@ export class MetadataSortList {
     );
 
     if (indexOf > -1) {
-      this.selectedCategories[indexOf].selected = false;
       this.selectedCategories.splice(indexOf, 1);
     }
   }
 
   replace(category: string, newCategory: string) {
-    const oldCategory = this.findByCategory(category);
-    oldCategory.selected = false;
-
     const newCategoryFound = this.findByCategory(newCategory);
-    newCategoryFound.selected = true;
 
     const indexOf = this.selectedCategories.findIndex(
       (metadataSort) => metadataSort.name === category
     );
 
     if (indexOf > -1) {
-      this.selectedCategories[indexOf] = newCategoryFound;
+      this.selectedCategories.splice(indexOf, 1, newCategoryFound);
     }
+  }
+
+  clear() {
+    this.selectedCategories = [];
   }
 
   toggleSort(category: string) {
     const found = this.findByCategory(category);
-    if (found) found.sort = found.sort === "asc" ? "desc" : "asc";
+    if (found) found.toggleSort();
   }
 
   convertToRouteParam(): string[] {
-    return this.selected.map((c) => `metadata.${c.name}:${c.sort}`);
+    return this.selected.map(
+      (c) => `${c.key}${c.name}${ORDER_BY_SEPARATOR}${c.sort}`
+    );
   }
 
   completeByRouteParams(sort: string) {
+    this.clear();
+
     if (!sort) return;
 
-    const sortParams = sort.replaceAll("metadata.", "").split(",");
+    const sortParams = sort.split(",");
 
     sortParams.forEach((sortParam) => {
-      const [name, sort] = sortParam.split(":");
+      const categories = sortParam.split(SORT_KEY_SEPARATOR);
+      const [name, sort] =
+        categories[categories.length - 1].split(ORDER_BY_SEPARATOR);
+
       const found = this.findByCategory(name);
+
       if (found) {
-        found.selected = true;
-        found.sort = sort as "asc" | "desc";
+        found.sort = sort as SortOptions;
 
         this.selectedCategories.push(found);
       }