+++ tests/factories.py
@@ -243,13 +243,15 @@ class MetadataPropertyFactory(BaseFactory):
     class Meta:
         model = MetadataProperty
 
+    # TODO: Remove this method and fix possible failing tests
     @classmethod
     def _create(cls, model_class, *args, **kwargs):
         default_settings = getattr(cls, "settings", {})
         settings = kwargs.get("settings", {})
         if settings:
-            default_settings.update(settings)
-            kwargs["settings"] = default_settings
+            new_settings = default_settings.copy()
+            new_settings.update(settings)
+            kwargs["settings"] = new_settings
         return super()._create(model_class, *args, **kwargs)
 
     name = factory.Sequence(lambda n: f"metadata-property-{n}")
@@ -276,13 +278,15 @@ class QuestionFactory(BaseFactory):
     class Meta:
         model = Question
 
+    # TODO: Remove this method and fix possible failing tests
     @classmethod
     def _create(cls, model_class, *args, **kwargs):
         default_settings = cls.settings.copy()
         settings = kwargs.get("settings", {})
         if settings:
-            default_settings.update(settings)
-            kwargs["settings"] = default_settings
+            new_settings = default_settings.copy()
+            new_settings.update(settings)
+            kwargs["settings"] = new_settings
         return super()._create(model_class, *args, **kwargs)
 
     name = factory.Sequence(lambda n: f"question-{n}")